requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-experimental>=0.0.50
tabulate>=0.9.0